% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_long.R
\docType{data}
\name{coronavirus_long}
\alias{coronavirus_long}
\title{Global COVID-19 daily data long form}
\format{
An object of class \code{data.frame} with 211184 rows and 4 columns.
}
\usage{
coronavirus_long
}
\description{
This package provides a long form dataset -- coronavirus_long, which was selected from the coronavirus dataset, contains the country, date, type of cases, and the number of cases.
}
\keyword{datasets}
