% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_wide.R
\docType{data}
\name{coronavirus_wide}
\alias{coronavirus_wide}
\title{Global COVID-19 daily data wide form}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50652 rows and 5 columns.
}
\usage{
coronavirus_wide
}
\description{
This package also contains a wide form dataset which contains date, country, confirmed, recovered and death.
}
\keyword{datasets}
