% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{run the whole shiny app which contains a datatable and a plotly plot}
\usage{
launch_app()
}
\description{
run the whole shiny app which contains a datatable and a plotly plot
}
