% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_countries.R
\name{n_countries}
\alias{n_countries}
\title{Count the number of countries}
\usage{
n_countries(data)
}
\arguments{
\item{data}{a data containing a country column}
}
\description{
Count the number of countries
}
