% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\docType{data}
\name{coronavirus}
\alias{coronavirus}
\title{Global COVID-19 daily data}
\format{
An object of class \code{data.frame} with 211184 rows and 7 columns.
}
\source{

}
\usage{
coronavirus
}
\description{
The coronavirus dataset is from the R package 'coronavirus'. There are 189 countries in this dataset.
}
\keyword{datasets}
