% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\docType{data}
\name{coronavirus}
\alias{coronavirus}
\title{Global COVID-19 daily data}
\format{
An object of class \code{data.frame} with 211184 rows and 7 columns.
}
\source{

}
\usage{
coronavirus
}
\description{
The coronavirus dataset is from the R package 'coronavirus'. There are 189 countries in this dataset.
This package provides a long form dataset -- coronavirus_long, which was selected from the coronavirus dataset, contains the country, date, type of cases, and the number of cases.
And this package also contains a wide form dataset which contains date, country, confirmed, recovered and death.
}
\keyword{datasets}
