% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdcovid19-package.R
\docType{package}
\name{cwdcovid19-package}
\alias{cwdcovid19}
\alias{cwdcovid19-package}
\title{cwdcovid19: Communicating With Data Course Assignment COVID19 Data Explore Package}
\description{
This package contains the shiny app which explores the data of COVID-19. This is an package for cwd course assignment. You could use this package to check the daily COVID-19 cases data of Chile or Bolivia using the launch_dt() function. Or use the launch_confirm_period() function to see the confirmed cases of your selected time period in these two country.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-1029YanMa}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-1029YanMa/issues}
}

}
\author{
\strong{Maintainer}: Yan Ma \email{ymaa0063@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
