% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coviddt.R
\docType{data}
\name{n_countries}
\alias{n_countries}
\title{Count the number of countries}
\format{
An object of class \code{function} of length 1.
}
\usage{
n_countries
}
\arguments{
\item{data}{a data containing a country column}
}
\description{
Count the number of countries
}
\keyword{datasets}
